requirements.txt
# This env built on python=3.10
pyyaml==6.0.2
ultralytics==8.3.165
transformers==4.53.2
shapely==2.1.1
rapidfuzz==3.13.0
fastapi==0.116.1
python-multipart==0.0.20
uvicorn==0.35.0